dev_v2/frontend/requirements.txt
google-genai>=1.0.0
PyJWT>=2.8.0
